+++ src/mcmc/hmc.jl
@@ -462,13 +462,15 @@ function AHMCAdaptor(::Hamiltonian, ::AHMC.AbstractMetric, nadapts::Int; kwargs.
 end
 
 """
-    post_sample_hook(chain::MCMCChains.Chains, sampler::Union{HMC,NUTS,HMCDA})
+    post_sample_hook(chain::MCMCChains.Chains, sampler::Union{HMC,NUTS,HMCDA}; verbose::Bool=true)
 
 Emit a warning message if there are divergent transitions in the chain.
 """
-function post_sample_hook(chain::MCMCChains.Chains, ::Union{HMC,NUTS,HMCDA})
+function post_sample_hook(
+    chain::MCMCChains.Chains, ::Union{HMC,NUTS,HMCDA}; verbose::Bool=true
+)
     n_divergent = round(Int, sum(skipmissing(vec(chain[:numerical_error]))))
-    if n_divergent > 0
+    if verbose && n_divergent > 0
         @warn "There were $n_divergent divergent transitions. Consider reparameterising your model or using a smaller step size. For adaptive samplers such as NUTS and HMCDA, consider increasing `target_accept`."
     end
     return nothing